+++ ui/pages/details/templates/bridge-details/bridge-details.tsx
